refactor: Extract getTxOptions and spending limit dispatch
