+++ src/utils/transactions.ts
@@ -2,6 +2,7 @@ import {
   ChainInfo,
   DateLabel,
   ExecutionInfo,
+  FEATURES,
   getTransactionDetails,
   MultisigExecutionDetails,
   MultisigExecutionInfo,
@@ -13,6 +14,9 @@ import { MetaTransactionData, OperationType } from '@gnosis.pm/safe-core-sdk-typ
 import { getGnosisSafeContractInstance } from '@/services/contracts/safeContracts'
 import extractTxInfo from '@/services/tx/extractTxInfo'
 import { createExistingTx } from '@/services/tx/txSender'
+import { AdvancedParameters } from '@/components/tx/AdvancedParams'
+import { TransactionOptions } from '@gnosis.pm/safe-core-sdk-types'
+import { hasFeature } from '@/utils/chains'
 
 export const makeTxFromDetails = (txDetails: TransactionDetails): Transaction => {
   const getMissingSigners = ({
@@ -132,4 +136,22 @@ export const getSafeTxs = (txs: TransactionDetails[], chainId: string, safeAddre
       return await createExistingTx(chainId, safeAddress, tx.txId, tx)
     }),
   )
+}
+
+export const getTxOptions = (params: AdvancedParameters, currentChain: ChainInfo | undefined): TransactionOptions => {
+  const txOptions = {
+    gasLimit: params.gasLimit?.toString(),
+    maxFeePerGas: params.maxFeePerGas?.toString(),
+    maxPriorityFeePerGas: params.maxPriorityFeePerGas?.toString(),
+    nonce: params.userNonce?.toString(),
+  } as TransactionOptions
+
+  // Some chains don't support EIP-1559 gas price params
+  if (currentChain && !hasFeature(currentChain, FEATURES.EIP1559)) {
+    txOptions.gasPrice = txOptions.maxFeePerGas
+    delete txOptions.maxFeePerGas
+    delete txOptions.maxPriorityFeePerGas
+  }
+
+  return txOptions
 }
